+++ crates/goose/src/acp/server/dispatch.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::providers::inventory::ensure_refresh_identity_current;
 
 impl HandleDispatchFrom<Client> for GooseAcpHandler {
     fn describe_chain(&self) -> impl std::fmt::Debug {
