+++ crates/goose/src/acp/server/manage_sessions.rs
